pyproject.toml
[tool.poetry]
name = "LLMQuiver"
version = "0.1.0"
description = "The auxiliary tool is used to invoke the online LLM service API, supporting local caching, prompt rendering, and configuration management."
authors = ["xrandx <kevtyle@hotmail.com>"]
readme = "README.md"
package-mode = false 

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
loguru = "^0.7.2"
tqdm = "^4.66.1"
openai = "^1.30.0"
tiktoken = "0.7.0"
toml = "^0.10.2"
pyyaml = "^6.0.2"
joblib = "^1.4.2"
filelock = "^3.15.4"

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
priority = "primary"
